+++ src/reactpy_router/components.py
@@ -1,18 +1,20 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Any
+from typing import TYPE_CHECKING, Any
 from urllib.parse import urljoin
 from uuid import uuid4
 
 from reactpy import component, html, use_connection
 from reactpy.backend.types import Location
-from reactpy.core.component import Component
-from reactpy.core.types import VdomDict
 from reactpy.web.module import export, module_from_file
 
 from reactpy_router.hooks import _use_route_state
 from reactpy_router.types import Route
+
+if TYPE_CHECKING:
+    from reactpy.core.component import Component
+    from reactpy.core.types import Key, VdomDict
 
 History = export(
     module_from_file("reactpy-router", file=Path(__file__).parent / "static" / "bundle.js"),
@@ -40,7 +42,7 @@ FirstLoad = export(
 link_js_content = (Path(__file__).parent / "static" / "link.js").read_text(encoding="utf-8")
 
 
-def link(attributes: dict[str, Any], *children: Any) -> Component:
+def link(attributes: dict[str, Any], *children: Any, key: Key | None = None) -> Component:
     """
     Create a link with the given attributes and children.
 
@@ -51,7 +53,7 @@ def link(attributes: dict[str, Any], *children: Any) -> Component:
     Returns:
         A link component with the specified attributes and children.
     """
-    return _link(attributes, *children)
+    return _link(attributes, *children, key=key)
 
 
 @component
@@ -68,7 +70,8 @@ def _link(attributes: dict[str, Any], *children: Any) -> VdomDict:
     if "href" in attributes and "to" not in attributes:
         attributes["to"] = attributes.pop("href")
     if "to" not in attributes:  # pragma: no cover
-        raise ValueError("The `to` attribute is required for the `Link` component.")
+        msg = "The `to` attribute is required for the `Link` component."
+        raise ValueError(msg)
     to = attributes.pop("to")
 
     attrs = {
@@ -132,7 +135,7 @@ def route(path: str, element: Any | None, *routes: Route) -> Route:
     return Route(path, element, routes)
 
 
-def navigate(to: str, replace: bool = False) -> Component:
+def navigate(to: str, replace: bool = False, key: Key | None = None) -> Component:
     """
     Navigate to a specified URL.
 
@@ -146,7 +149,7 @@ def navigate(to: str, replace: bool = False) -> Component:
     Returns:
         The component responsible for navigation.
     """
-    return _navigate(to, replace)
+    return _navigate(to, replace, key=key)
 
 
 @component