+++ tests/test_router.py
@@ -136,11 +136,12 @@ async def test_use_params(display: DisplayFixture):
 
     await display.show(sample)
 
-    for path, expected_params in [
+    for path, _expected_params in [
         ("/first/1", {"first": "1"}),
         ("/first/1/second/2", {"first": "1", "second": "2"}),
         ("/first/1/second/2/third/3", {"first": "1", "second": "2", "third": "3"}),
     ]:
+        expected_params = _expected_params
         await display.goto(path)
         await display.page.wait_for_selector("#success")